fix(tform): potential data race on identityofthreads
During the initialization of worker threads, Helgrind detects a
potential data race on the identityofthreads variable. This arises
because SetIdentity() increases identityofthreads for each thread, while
WhoAmI(), invoked within InitializeOneThread(), checks the value of
identityofthreads without any lock. In practice, this is OK for commonly
used platforms, but the C language does not guarantee atomic access to
int variables.

2 possible solutions to address this issue:

(1) Lock numberofthreadslock during the reading of identityofthreads in
WhoAmI(). However, since WhoAmI() is called in many places, this
approach may lead to a noticeable performance reduction.

(2) Put a barrier between the calls of SetIdentity() and
InitializeOneThread() to ensure that WhoAmI() and SetIdentity() are not
executed simultaneously.

This patch implements the latter solution.