test(check.rb): --stat details before summary
Previously, the detailed output from the --stat option appeared after
the test suite summary, making it harder to access the summary when the
details were lengthy. This patch tries to display the --stat details
before the summary, which should work with test-unit >= 3.0.9.

It also tries to use the "failure" colour for the failed test lines in
the details.