Warn by default if parallel execution is vetoed due to dollar variables
Before this commit FORM only warns when disabling parallel execution in
a module if `on allwarnings` is set. This makes sense, since for most of
the causes it is not straightforward to modify the code in a way that
re-enables parallel execution. However, modifying a dollar variable without
a corresponding `moduleoption` statement is often a trivial bug and the
user should be warned by default.