feat(grcc): add QGRAF options
This patch implements the following QGRAF options:

    onepi/onepr
    onshell/offshell
    nosigma/sigma
    nosnail/snail
    notadpole/tadpole
    simple/notsimple
    bipart/nonbipart
    cycli/cyclr
    floop

The following options are not implemented:

    topol
    onevi/onevr
    onshellx/offshellx

The default options are updated to generate the set of all connected
graphs. Specifying additional options can generate a subset of the
default output.

Example:

    #define LOOPS "2"
    Vectors Q1,Q2,p1,...,p8;
    Set QQ:Q1,Q2;
    Set PP:p1,...,p8;
    Local F = topologies_(`LOOPS',2,{3,4},QQ,PP);
    Print +ss;
    .end

This generates 24 terms, corresponding to 24 topologies, due to the
updated default options. When applied to the 3-loop case, the same
example yields 240 terms.

Note: The above example crashes when called with
"`LOOPS',-2,{3,4},QQ,PP". This behaviour was also seen in previous
versions.