fix: use a scratch buffer in simp3btoken
Issue #251 crashes because "fill" overtakes "s" and produces malformed tokens,
leading to an infinite loop. Avoid this possibility by writing the output
tokens into a scratch buffer, and copy the result back at the end.

Additionally, re-allocate the buffers if fill is "close to" the end of the
scratch buffer. Here "close to" is a guess at how much space might be required
per iteration.